Remove the special case of passing pydata in tests
The strict treatment of the passed pydata type broke many tests where a
memory drop is initialised with data. It also broke the dask_emulation
since that is bypassing the normal treatment and also supports passing
complex objects as default arguments. The tests and dask_emulation
are now more aligned with the complete definition of graph nodes
through EAGLE.
