+++ daliuge-engine/dlg/apps/pyfunc.py
@@ -64,31 +64,6 @@ logger = logging.getLogger(f"dlg.{__name__}")
 
 MAX_IMPORT_RECURSION = 100
 
-def serialize_func(f, serialize=True):
-    if isinstance(f, str):
-        parts = f.split(".")
-        f = getattr(importlib.import_module(".".join(parts[:-1])), parts[-1])
-
-    fser = base64.b64encode(dill.dumps(f)).decode() if serialize else f
-    # fser = inspect.getsource(f)
-    fdefaults = {"args": [], "kwargs": {}}
-    adefaults = {"args": [], "kwargs": {}}
-    a = inspect.getfullargspec(f)
-    if a.defaults:
-        fdefaults["kwargs"] = dict(
-            zip(
-                a.args[-len(a.defaults):],
-                [serialize_data(d) for d in a.defaults],
-            )
-        )
-        adefaults["kwargs"] = dict(
-            zip(a.args[-len(a.defaults):], [d for d in a.defaults])
-        )
-    logger.debug("Introspection of function %s: %s", f, a)
-    logger.debug("Defaults for function %r: %r", f, adefaults)
-    return fser, fdefaults
-
-
 def import_using_name(app, fname, curr_depth):
     if curr_depth > MAX_IMPORT_RECURSION:
         raise InvalidDropException(
